+++ Patches/LoadPatches.cs
@@ -1,6 +1,5 @@
 using HarmonyLib;
 using MoreCounterplay.Behaviours;
-using MoreCounterplay.Config;
 using MoreCounterplay.Items;
 using MoreCounterplay.Util;
 using UnityEngine;
@@ -8,37 +7,66 @@ using UnityEngine;
 namespace MoreCounterplay.Patches
 {
     /// <summary>
-    ///     Patch for loading texture assets upon loading in.
+    ///     Patches for loading/unloading stuff upon joining or leaving a lobby.
     /// </summary>
-    [HarmonyPatch(typeof(StartOfRound))]
-    internal class StartOfRoundPatch
+    [HarmonyPatch]
+    internal class LoadPatches
     {
-        [HarmonyPatch(typeof(StartOfRound), nameof(StartOfRound.Start))]
+        [HarmonyPatch(typeof(GameNetworkManager), nameof(GameNetworkManager.StartDisconnect))]
         [HarmonyPostfix]
-        private static void FindAndLoadTextures()
+        private static void OnDisconnect()
         {
+            // Remove added components from the Coilhead prefab.
+            CoilExplosion.RadioactiveFirePrefab?.transform.SetParent(null);
+            Object.Destroy(CoilExplosion.CoilheadPrefab?.GetComponent<CoilExplosion>());
+
+            // Reset Coilhead prefab to vanilla values.
+            if (CoilExplosion.CoilheadPrefab?.TryGetComponent(out SpringManAI coilhead) ?? false)
+            {
+                coilhead.enemyHP = 3;
+                coilhead.enemyType.canDie = false;
+            }
+        }
+
+        /// <summary>
+        ///     Finds vanilla prefabs and grabs their textures or adds components to them.
+        /// </summary>
+        internal static void FindAndModifyPrefabs()
+        {
+            // Currently, only the Coilhead counterplay requires textures loaded at runtime.
+            if (!MoreCounterplay.Settings.EnableCoilheadCounterplay) return;
+
             MoreCounterplay.Log("Loading prefab textures...");
 
             // Try to find and obtain the Coilhead's enemy prefab.
             if (VanillaPrefabUtils.GetInsideEnemyPrefab("Spring", out GameObject? coilheadPrefab))
             {
+                // Save reference to vanilla Coilhead prefab for convenience.
+                CoilExplosion.CoilheadPrefab = coilheadPrefab;
+
                 if (HeadItem.Prefab != null && coilheadPrefab != null)
                 {
+                    // Modify Coilhead health to the configured amount and allow it to die.
                     SpringManAI coilhead = coilheadPrefab.GetComponent<SpringManAI>();
-                    coilhead.enemyHP = ConfigSettings.SpringDurability.Value;
+                    coilhead.enemyHP = MoreCounterplay.Settings.SpringDurability;
                     coilhead.enemyType.canDie = true;
 
-                    // Add explosion network behaviour script to the Coilhead prefab.
-                    CoilExplosion coilExplosion = coilheadPrefab.AddComponent<CoilExplosion>();
-                    coilExplosion.enabled = false;
+                    if (MoreCounterplay.Settings.LoreAccurateCoilheads)
+                    {
+                        // Add explosion network behaviour script to the Coilhead prefab.
+                        coilheadPrefab.AddComponent<CoilExplosion>().enabled = false;
 
-                    // Add radioactive fire particle effects container to the Coilhead prefab.
-                    CoilExplosion.RadioactiveFirePrefab?.transform.SetParent(coilheadPrefab.transform.Find("SpringManModel"), false);
-                    CoilExplosion.RadioactiveFirePrefab?.SetActive(false);
+                        // Add radioactive fire particle effects container to the Coilhead prefab.
+                        CoilExplosion.RadioactiveFirePrefab?.transform.SetParent(coilheadPrefab.transform.Find("SpringManModel"), false);
+                        CoilExplosion.RadioactiveFirePrefab?.SetActive(false);
+                    }
 
-                    // Obtain Coilhead material from its enemy prefab and assign it to the 'Coilless Coilhead' scrap item prefab.
-                    Material coilheadMaterial = coilheadPrefab.transform.Find("SpringManModel/Head").GetComponent<MeshRenderer>().material;
-                    HeadItem.Prefab.GetComponent<MeshRenderer>().material = coilheadMaterial;
+                    if (MoreCounterplay.Settings.DropHeadAsScrap)
+                    {
+                        // Obtain Coilhead material from its enemy prefab and assign it to the 'Coilless Coilhead' scrap item prefab.
+                        Material coilheadMaterial = coilheadPrefab.transform.Find("SpringManModel/Head").GetComponent<MeshRenderer>().material;
+                        HeadItem.Prefab.GetComponent<MeshRenderer>().material = coilheadMaterial;
+                    }
                 }
                 else
                 {
@@ -48,7 +76,7 @@ namespace MoreCounterplay.Patches
             }
 
             // Try to find and obtain the Forest Giant enemy prefab.
-            if (VanillaPrefabUtils.GetOutsideEnemyPrefab("ForestGiant", out GameObject? giantPrefab))
+            if (MoreCounterplay.Settings.LoreAccurateCoilheads && VanillaPrefabUtils.GetOutsideEnemyPrefab("ForestGiant", out GameObject? giantPrefab))
             {
                 if (CoilExplosion.RadioactiveFirePrefab != null && giantPrefab != null)
                 {
@@ -64,10 +92,6 @@ namespace MoreCounterplay.Patches
             }
 
             MoreCounterplay.Log("Finished loading textures!");
-
-            // Unpatch this method after running once.
-            MoreCounterplay.Harmony?.Unpatch(AccessTools.Method(typeof(StartOfRound), nameof(StartOfRound.Start)),
-                HarmonyPatchType.Postfix, MyPluginInfo.PLUGIN_GUID);
         }
     }
 }